+++ NEWS.md
@@ -1,5 +1,7 @@
 # tidyr 0.8.2.9000
 
+* `crossing()` preserves factor levels (#410).
+
 # tidyr 0.8.2
 
 * `separate()` now accepts `NA` as a column name in the `into` argument to